% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuner.R
\name{apply_tuned_clustering}
\alias{apply_tuned_clustering}
\title{Apply chosen k/res to an object (returns updated object)}
\usage{
apply_tuned_clustering(
  obj,
  dims_use,
  k_best,
  res_best,
  prune = 0.1,
  algorithm = 4,
  seed = 1
)
}
\arguments{
\item{obj}{Seurat object to update}

\item{dims_use}{Integer vector of PC indices used to build the graph}

\item{k_best}{Integer; chosen k.param}

\item{res_best}{Numeric; chosen resolution}

\item{prune}{Numeric; prune.SNN passed to FindNeighbors}

\item{algorithm}{Integer; clustering algorithm (e.g., 4 = Leiden)}

\item{seed}{Integer random seed}
}
\value{
Updated Seurat object
}
\description{
Apply chosen k/res to an object (returns updated object)
}
