% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{default_stress_genes}
\alias{default_stress_genes}
\title{Default stress genes}
\usage{
default_stress_genes(species = c("mouse", "human"))
}
\arguments{
\item{species}{'mouse' or 'human'}
}
\value{
Character vector of stress-related genes
}
\description{
Default stress genes
}
