% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.add_module_score_safe}
\alias{.add_module_score_safe}
\title{Safe wrapper around AddModuleScore that returns the appended column names}
\usage{
.add_module_score_safe(
  obj,
  features,
  assay,
  name = "MS__",
  nbin = 24,
  ctrl = 100,
  debug_dir = NULL
)
}
\description{
Safe wrapper around AddModuleScore that returns the appended column names
}
\keyword{internal}
