% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{.ensure_visual_embeddings}
\alias{.ensure_visual_embeddings}
\title{Ensure PCA/UMAP are present for nice plots (Seurat v5)}
\usage{
.ensure_visual_embeddings(
  seu,
  assay = "RNA",
  pc_dims = 1:20,
  debug = getOption("scQCenrich.debug", FALSE)
)
}
\description{
Ensure PCA/UMAP are present for nice plots (Seurat v5)
}
\keyword{internal}
