% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{.norm_cells}
\alias{.norm_cells}
\title{Normalize cell barcodes in several safe modes (used for cross-obj alignment)}
\usage{
.norm_cells(cells, mode = c("basic", "core16"))
}
\arguments{
\item{cells}{character vector of cell names}

\item{mode}{"basic" strips sample prefix (before ':'), strips trailing 'x',
and strips trailing '-\if{html}{\out{<digits>}}' (e.g., '-1'); "core16" returns
just the last 16 A/C/G/T characters if present.}
}
\description{
Normalize cell barcodes in several safe modes (used for cross-obj alignment)
}
\keyword{internal}
