% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures_generic.R
\name{generic_signatures}
\alias{generic_signatures}
\title{Generic, cross-tissue marker signatures (broad/extended)}
\usage{
generic_signatures(
  species = c("mouse", "human"),
  tier = c("broad", "extended")
)
}
\arguments{
\item{species}{"mouse" or "human"}

\item{tier}{"broad" or "extended"}
}
\description{
Generic, cross-tissue marker signatures (broad/extended)
}
