% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{qc_featuremaps}
\alias{qc_featuremaps}
\title{Make & save UMAP featuremaps for QC metrics}
\usage{
qc_featuremaps(
  obj,
  metrics = c("nCount_QC", "nFeature_QC", "pctMT_QC", "MALAT1_frac_QC",
    "stress_score_QC", "intronic_frac_QC", "spliced_frac_QC", "unspliced_frac_QC",
    "u2s_ratio_QC", "dbl_score"),
  save_dir = "qc_outputs",
  prefix = "qc",
  ncol = 4,
  logify = c(nCount_QC = TRUE, nFeature_QC = TRUE, u2s_ratio_QC = TRUE, dbl_score =
    FALSE)
)
}
\arguments{
\item{obj}{Seurat object (must have UMAP). Use .ensure_pca_umap() beforehand.}

\item{metrics}{Vector of meta column names to plot (only those present are used)}

\item{save_dir}{Where to write PNGs}

\item{prefix}{File prefix for images}

\item{ncol}{Number of columns in the grid image}

\item{logify}{Named logical vector: which metrics to log1p/ log10; defaults below}
}
\description{
Make & save UMAP featuremaps for QC metrics
}
