% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{render_qc_report}
\alias{render_qc_report}
\title{Render a pretty HTML QC report}
\usage{
render_qc_report(
  obj,
  metrics,
  status_df,
  save_dir = "qc_outputs",
  report_file = file.path(save_dir, "qc_report.html"),
  theme_bootswatch = "minty",
  self_contained = TRUE,
  sample_col = NULL,
  template = NULL,
  species = c("mouse", "human")
)
}
\arguments{
\item{obj}{Seurat object (with qc_status in meta)}

\item{metrics}{data.frame from calcQCmetrics()}

\item{status_df}{data.frame from flagLowQuality()}

\item{save_dir}{output dir for assets/plots (and where the Rmd will be copied)}

\item{report_file}{output HTML path (default: file.path(save_dir, "qc_report.html"))}

\item{theme_bootswatch}{Bootswatch theme name}

\item{self_contained}{embed all assets into single HTML (TRUE)}

\item{sample_col}{optional meta column for per-sample summaries}

\item{template}{path to a .Rmd template; if NULL, copy bundled template to save_dir and use it}

\item{species}{Character scalar, one of "mouse" or "human"}
}
\description{
Render a pretty HTML QC report
}
