% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-globals.R
\docType{data}
\name{toy_seu}
\alias{toy_seu}
\title{toy_seu}
\format{
A \code{Seurat} object with minimal counts and metadata.
}
\source{
Generated in package development for demos.
}
\usage{
data(toy_seu)
}
\description{
A small example Seurat object used in examples and tests.
}
\keyword{datasets}
